[large tensor] Use int64_t instead of int/int32_t for dims (#76290)
* add new enforce

* apply int32-dims

* fix

* add rule

* fix

* fix rule

* remove  PADDLE_ENFORCE_LE_INT_MAX

* fix

* fix ast-grep

+++ paddle/phi/kernels/funcs/vol2col.cu
@@ -135,12 +135,29 @@ void Vol2ColFunctor<DeviceContext, T>::operator()(
       (data_layout != DataLayout::kNHWC ? vol.dims()[2] : vol.dims()[1]);
   int input_width =
       (data_layout != DataLayout::kNHWC ? vol.dims()[3] : vol.dims()[2]);
-  int filter_depth = col->dims()[1];
-  int filter_height = col->dims()[2];
-  int filter_width = col->dims()[3];
-  int output_depth = col->dims()[4];
-  int output_height = col->dims()[5];
-  int output_width = col->dims()[6];
+  int64_t filter_depth = col->dims()[1];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
+
+  int64_t filter_height = col->dims()[2];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
+
+  int64_t filter_width = col->dims()[3];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
+
+  int64_t output_depth = col->dims()[4];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
+
+  int64_t output_height = col->dims()[5];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
+
+  int64_t output_width = col->dims()[6];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
 
   bool paddings_size_is_6 = (paddings.size() == 6);
   int pad_d_forth = paddings_size_is_6 ? paddings[0] : paddings[0];
@@ -344,12 +361,29 @@ void Col2VolFunctor<DeviceContext, T>::operator()(
       (data_layout != DataLayout::kNHWC ? vol->dims()[2] : vol->dims()[1]);
   int input_width =
       (data_layout != DataLayout::kNHWC ? vol->dims()[3] : vol->dims()[2]);
-  int filter_depth = col.dims()[1];
-  int filter_height = col.dims()[2];
-  int filter_width = col.dims()[3];
-  int output_depth = col.dims()[4];
-  int output_height = col.dims()[5];
-  int output_width = col.dims()[6];
+  int64_t filter_depth = col.dims()[1];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
+
+  int64_t filter_height = col.dims()[2];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
+
+  int64_t filter_width = col.dims()[3];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
+
+  int64_t output_depth = col.dims()[4];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
+
+  int64_t output_height = col.dims()[5];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
+
+  int64_t output_width = col.dims()[6];
+  // TODO(large-tensor): downstream functors may still use int; guard until
+  // upgraded.
 
   bool paddings_size_is_6 = (paddings.size() == 6);
   int pad_d_forth = paddings_size_is_6 ? paddings[0] : paddings[0];
